moving GridClient out ImageDateMerger as it is getting complex
